requirements.txt
python-dotenv==1.0.0
pypdf==3.17.1
langchain-community==0.0.29
langchain==0.1.13
langchain-openai==0.0.8
faiss-cpu==1.8.0
openai==1.12.0
sentence-transformers==2.3.1
httpx==0.27.2
Flask==3.0.0
gunicorn==21.2.0
fireworks-ai==0.15.4
